requirements.txt
Flask-APScheduler
alpaca-trade-api
alpaca-py
bs4
cachetools
flask
GitPython
matplotlib
numpy
pandas
polygon-api-client
psycopg2
pytest
pytest-mock
python-dateutil
pytz
ratelimiter
recordclass
requests
retrying
sqlalchemy
tabulate
tqdm
waitress
